派工和加工顺序的一些更改

+++ src/main/java/com/ruoyi/project/pipe/workPipe/domain/WorkPipe.java
@@ -84,7 +84,7 @@ public class WorkPipe extends BaseEntity
 		this.shapeShipId = shapeShipId;
 	}
 
-	public double getShapeShipId()
+	public Integer getShapeShipId()
 	{
 		return shapeShipId;
 	}
